feat: 말씀카드 생성 UX 반영
