+++ src/pages/BibleCardPage/BibleCardCarousel.tsx
@@ -1,15 +1,10 @@
 import { useState, useEffect } from "react";
-import { cn } from "@/lib/utils";
 import BibleCard1 from "@/assets/bibleCard/BibleCard1.jpeg";
 import BibleCard2 from "@/assets/bibleCard/BibleCard2.jpeg";
 import BibleCard3 from "@/assets/bibleCard/BibleCard3.jpeg";
 import BibleCard4 from "@/assets/bibleCard/BibleCard4.jpeg";
 
-interface BibleCardCarouselProps {
-  className?: string;
-}
-
-const BibleCardCarousel: React.FC<BibleCardCarouselProps> = ({ className }) => {
+const BibleCardCarousel: React.FC = () => {
   const images = [BibleCard1, BibleCard2, BibleCard3, BibleCard4];
 
   const [currentIndex, setCurrentIndex] = useState(0);
@@ -23,18 +18,14 @@ const BibleCardCarousel: React.FC<BibleCardCarouselProps> = ({ className }) => {
   }, [images.length]);
 
   return (
-    <div
-      className={cn(className, "relative w-full overflow-hidden rounded-lg")}
-    >
+    <div className="overflow-hidden rounded-lg shadow-lg">
       <div
         className="flex transition-transform duration-500 ease-in-out"
         style={{ transform: `translateX(-${currentIndex * 100}%)` }}
         aria-live="polite"
       >
         {images.map((src, index) => (
-          <div key={index} className="w-full flex-shrink-0">
-            <img src={src} className="w-full aspect-[4/6] border" />
-          </div>
+          <img key={index} src={src} className="w-full  aspect-auto" />
         ))}
       </div>
     </div>
